Add support for default input object values
It is not clear that this is the right approach: we convert the default values from the query to rust literals for the input objects. The default value is then sent twice: in the query string and in the variable.

The reason for using this approach is that Rust structs do not support only setting some fields and not others, so there is no distinction between absent and null fields: there is no way to say if the field in the Variables struct is explicitly set to null or left out (expecting the default value to be used).

Maybe we could generate something other than an Option here, but we would have to figure out how to make this ergonomic.