graphql_client_codegen: fix normalization
Normalization as a feature means that any usage of it within a
dependency tree turns it on for all users in the tree. Since the feature
changes code generation and requires changes to use properly, this
should instead be controlled by the callsite generating the code. For
this, we have attributes available.

Fixes: #283