Always sort scalar types by name when returning them
Previously these were returned in an essentially random order. This would lead
to flapping in the generated code when running the client CLI against the same
schema multiple times.

In the application I'm writing, I check in the generated schema, so this sort
of flapping in the generated code creates unwanted noise in the commit
history.